feat: remove improved time period filter from feature flag (#4645)
